[CF-120] Fix "Port is still in use" when attempt boot VM second time
Fix an issue when we trying boot VM in the same availability zone,
but on another host. For example we have this case when the VM failed
during deployment on the first host. Before this patch we had exception
"Port is still in use", because port is still attached to the failed VM.
With this patch CloudFerry recreate ports for the VM and trying boot it
on another host.